drivers: eip76d_trng: move into core/drivers
Commit 5b6c7df764e5 ("plat-k3: drivers: Refactor SA2UL driver")
made the eip76_trng driver independent of SA2UL/SA3UL.

Move the driver from core/arch/arm/plat-k3 into core/drivers,
so that it can be used on other platforms.
