+++ core/drivers/sub.mk
@@ -23,6 +23,7 @@ srcs-$(CFG_HI16XX_RNG) += hi16xx_rng.c
 srcs-$(CFG_LPC_UART) += lpc_uart.c
 srcs-$(CFG_SCIF) += scif.c
 srcs-$(CFG_DRA7_RNG) += dra7_rng.c
+srcs-$(CFG_EIP76D_TRNG) += eip76d_trng.c
 srcs-$(CFG_STIH_UART) += stih_asc.c
 srcs-$(CFG_ATMEL_UART) += atmel_uart.c
 srcs-$(CFG_ATMEL_TRNG) += atmel_trng.c
