+++ core/include/drivers/eip76d_trng.h
